Cleanup and refactored some code to fix a bug
Cleanup a  couple of methods to remove the [Done] button only when the
user clicks on the [Convert] button. This fixes the issue were if a
text field is selected and then the user switches to another text field
the [Done] button will disappear and they keyboard couldn't be
dismissed.